btConvexHull: Fix warnings/fpermissive

+++ src/btConvexHull/btAlignedAllocator.cpp
@@ -64,11 +64,11 @@ static inline void* btAlignedAllocDefault(size_t size, int32_t alignment)
 {
     void* ret;
     char* real;
-    unsigned long offset;
+    uintptr_t offset;
 
     real = (char*)sAllocFunc(size + sizeof(void*) + (alignment - 1));
     if (real) {
-        offset = (alignment - (unsigned long)(real + sizeof(void*))) & (alignment - 1);
+        offset = (alignment - (uintptr_t)(real + sizeof(void*))) & (alignment - 1);
         ret = (void*)((real + sizeof(void*)) + offset);
         *((void**)(ret)-1) = (void*)(real);
     }
